+++ app/renderer/js/main.ts
@@ -22,6 +22,7 @@ import Logger = require('./utils/logger-util');
 import CommonUtil = require('./utils/common-util');
 import EnterpriseUtil = require('./utils/enterprise-util');
 import Messages = require('./../../resources/messages');
+import { notificationHandler } from './notification/winrt-helper';
 
 interface FunctionalTabProps {
 	name: string;
@@ -953,8 +954,26 @@ class ServerManagerView {
 			});
 		});
 
-		ipcRenderer.on('notifCall', (event: Event, notifObject: object) => {
+		interface NotifObject{
+			url: string,
+			tag: string,
+			response: string
+		};
+		ipcRenderer.on('notifCall', (event: Event, notifObject: NotifObject) => {
 			console.log(notifObject);
+			let url = notifObject.url;
+			if (url.charAt(url.length - 1) === '/') {
+				// remove / character at the end of the url
+				url = url.slice(0, -1);
+			}
+			console.log(url);
+			for (let tab=0; tab<this.tabs.length; ++tab) {
+				if (url.startsWith(this.tabs[tab].webview.props.url)) {
+					console.log(this.tabs[tab].webview.props.url);
+					this.activateTab(tab);
+				}
+			}
+			notificationHandler(notifObject.response, notifObject.tag);
 		});
 
 		ipcRenderer.on('render-taskbar-icon', (event: Event, messageCount: number) => {
